backward works

+++ extra/lr_scheduler.py
@@ -68,8 +68,8 @@ class CosineAnnealingLR(LR_Scheduler):
 class OneCycleLR(LR_Scheduler):
   def __init__(self, optimizer: Optimizer, max_lr: float, div_factor: float, final_div_factor: float, total_steps: int, pct_start: float,
                anneal_strategy: str = 'linear', cycle_momentum: bool = False):
-    self.initial_lr = Tensor([max_lr / div_factor])
-    self.max_lr = Tensor([max_lr])
+    self.initial_lr = Tensor([max_lr / div_factor], dtype=dtypes.float32)
+    self.max_lr = Tensor([max_lr], dtype=dtypes.float32)
     self.min_lr = self.initial_lr/final_div_factor
     super().__init__(optimizer)
     self.total_steps = total_steps
